+++ src/tools/Built-in/Command_tools.ts
@@ -1,10 +1,13 @@
 import * as vscode from 'vscode';
+import { ECommandOption, getCommandOptions } from '../../configUI';
+import { enumLog } from '../enumErr';
+import { getAllFunc } from '../Func/getAllFunc';
 import type { TCommandElement } from './Command';
 import { LineCommand } from './Command';
+import { CSnippetCommand } from './CSnippetCommand';
 
 function commandElement2Md(Element: TCommandElement): vscode.MarkdownString {
     const {
-        keyRawName,
         body,
         doc,
         link, // string | undefined
@@ -12,55 +15,62 @@ function commandElement2Md(Element: TCommandElement): vscode.MarkdownString {
     } = Element;
     const md: vscode.MarkdownString = new vscode.MarkdownString('', true)
         .appendMarkdown('Command')
-        .appendCodeblock(keyRawName, 'ahk')
         .appendCodeblock(body, 'ahk')
         .appendMarkdown(doc)
         .appendMarkdown('\n')
-        .appendMarkdown(`[(Read Doc)](${link ?? 'https://www.autohotkey.com/docs/commands/index.htm'})`) // FIXME
-        .appendMarkdown('\n\n***')
-        .appendMarkdown('\n\n*exp:*');
+        .appendMarkdown(`[(Read Doc)](${link ?? 'https://www.autohotkey.com/docs/commands/index.htm'})`); // FIXME
+
     // .appendCodeblock(exp.join('\n'));
     if (exp !== undefined && exp.length > 0) {
-        md.appendCodeblock(exp.join('\n'), 'ahk');
+        md
+            .appendMarkdown('\n\n***')
+            .appendMarkdown('\n\n*exp:*')
+            .appendCodeblock(exp.join('\n'), 'ahk');
     }
     md.supportHtml = true;
     return md;
 }
 
 type TCommandMDMap = ReadonlyMap<string, vscode.MarkdownString>;
-type TSnippetCommand = readonly vscode.CompletionItem[];
+type TSnippetCommand = readonly CSnippetCommand[];
 
 export const [snippetCommand, CommandMDMap] = ((): [TSnippetCommand, TCommandMDMap] => {
     const map1 = new Map<string, vscode.MarkdownString>();
-    const tempList: vscode.CompletionItem[] = [];
+    const tempList: CSnippetCommand[] = [];
     for (const [k, v] of Object.entries(LineCommand)) {
-        const md = commandElement2Md(v);
+        const md: vscode.MarkdownString = commandElement2Md(v);
         map1.set(k, md);
 
-        const { keyRawName, body, recommended } = v;
-
-        if (recommended !== undefined && !recommended) continue;
-
-        const item: vscode.CompletionItem = new vscode.CompletionItem({
-            label: keyRawName,
-            description: 'Command',
-        });
-        item.kind = vscode.CompletionItemKind.Field; // icon of https://code.visualstudio.com/docs/editor/intellisense#_types-of-completions
-        item.insertText = new vscode.SnippetString(body);
-        item.detail = 'Command of AHK (neko-help)'; // description
-        item.documentation = md;
-
-        tempList.push(item);
+        tempList.push(new CSnippetCommand(k, v, md));
     }
     return [tempList, map1];
 })();
 
-export function getSnippetCommand(lStr: string, position: vscode.Position): readonly vscode.CompletionItem[] {
-    const subStr = lStr.slice(0, position.character).trim();
+export function getSnippetCommand(lStr: string, position: vscode.Position): TSnippetCommand {
+    const subStr: string = lStr.slice(0, position.character).trim();
 
-    return (/^\w*$/ui).test(subStr)
-        ? snippetCommand
-        : [];
+    if (!(/^\w*$/ui).test(subStr)) return [];
+    //
+
+    const opt: ECommandOption = getCommandOptions();
+    switch (opt) {
+        case ECommandOption.All:
+            return snippetCommand;
+
+        case ECommandOption.Recommended:
+            return snippetCommand.filter((v) => v.recommended);
+
+        case ECommandOption.noSameFunc: {
+            const fnMap = getAllFunc();
+            return snippetCommand.filter((v) => !fnMap.has(v.upName));
+        }
+        case ECommandOption.notProvided:
+            return [];
+
+        default:
+            enumLog(opt);
+            return [];
+    }
 }
 
 export function getHoverCommand(